#6 replace shorts with ints

+++ src/main/java/org/jmasonry/jvm/classfile/ClassFileVersion.java
@@ -5,12 +5,12 @@ import org.jmasonry.jvm.compiler.CompilationUnitBuilder;
 public enum ClassFileVersion {
     JAVA_10(53, 0);
 
-    private final short major;
-    private final short minor;
+    private final int major;
+    private final int minor;
 
     ClassFileVersion(int major, int minor) {
-        this.major = (short) major;
-        this.minor = (short) minor;
+        this.major = major;
+        this.minor = minor;
     }
 
     public void write(CompilationUnitBuilder builder) {
